Правки по проекту

+++ src/test/java/tests/petstore/tests/api/steps/PetSteps.java
@@ -36,36 +36,36 @@ public class PetSteps {
 
     @Step("Обновление питомца")
     public void updatePet(String name, String status) {
-        updatedPetResponse = api.pet.updatePet(petResponse.getId(), name, status);
+        updatedPetResponse = api.pet.updatePet(petResponse.getId().intValue(), name, status);
     }
 
     @Step("Получение питомца по id")
     public PetResponseModel getPet() {
-        return api.pet.getPet(petResponse.getId());
+        return api.pet.getPet(petResponse.getId().intValue());
     }
 
     @Step("Удаление питомца")
     public int deletePet() {
-        return api.pet.deletePet(petResponse.getId());
+        return api.pet.deletePet(petResponse.getId().intValue());
     }
 
     @Step("Удаление питомца по id")
-    public int deletePet(Long id) {
+    public int deletePet(int id) {
         return api.pet.deletePet(id);
     }
 
     @Step("Обновление несуществующего питомца")
-    public PetNotFoundResponseModel updatePetNotFound(Long id, String name, String status) {
+    public PetNotFoundResponseModel updatePetNotFound(int id, String name, String status) {
         return api.pet.updatePetNotFound(id, name, status);
     }
 
     @Step("Поиск удаленного или несуществующего питомца по id")
     public PetNotFoundResponseModel getPetNotFound() {
-        return api.pet.getPetNotFound(petResponse.getId());
+        return api.pet.getPetNotFound(petResponse.getId().intValue());
     }
 
     @Step("Поиск несуществующего питомца по id")
-    public PetNotFoundResponseModel getPetNotFound(Long id) {
+    public PetNotFoundResponseModel getPetNotFound(int id) {
         return api.pet.getPetNotFound(id);
     }
 }
